+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!--## Unreleased-->
 
+## [2.0.1] - 2017-05-15
+
+* Many errors were changed to warnings so they wouldn't stop builds from completing.
+
 ## [2.0.0] - 2017-05-15
 
 * Stable release of the Polymer Analyzer.